Moved server from internal to cmd
