+++ services/Hangout/cmd/router.go
@@ -1,4 +1,4 @@
-package server
+package cmd
 
 import (
 	"net/http"
